crews/stock_analysis/pyproject.toml
[project]
name = "stock_analysis"
version = "0.1.0"
description = ""
authors = [{name = "Your Name", email = "you@example.com"}]
readme = "README.md"
requires-python = ">=3.12,<=3.13"
dependencies = [
    "crewai[tools]>=0.152.0",
    "python-dotenv>=1.0.1",
    "html2text>=2024.2.26",
    "sec-api>=1.0.20",
    "boto3>=1.37.51",
    "opentelemetry-api>=1.36.0",
    "opentelemetry-sdk>=1.36.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.36.0",
    "openinference-instrumentation-crewai>=0.1.14",
]

[project.scripts]
stock_analysis = "stock_analysis.main:run"
train = "stock_analysis.main:train"
